fix: apply drift correction to rtc isr
Since we use real-time counters instead of real-time clocks, the
resolution of the timer is not as we need for a smart-watch. To
investigate it, we made several experiments and saw that the drift is
0.06 seconds per second.

Within this patch, we add a very simple drift correction mechanism where
we add extra seconnds after a known period of time. The current
implementation adds 4 seconds per minute. (4/60 = 0.06)